Changed shadow camera FOV to remove shadow anomaly.
Because the shadow camera of a single light only covered part of the table
and not all of it, the shadow disappeared with a clipping when the shadow
exited the camera frustrum. Perhaps it's possible to make the shadow fade
out instead of increasing the FOV like I did here?